Moved the ElevationOverlay class to its own file so we don't have to hide it. Small updates from code review feedback.
